Improved UX, code cleanup
